+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Sort by generation on `student-index`.
+- Changed from 'Hasil' to 'Delegasi'.
+
 # 1.0.0
 
 - Added UI.